third_party/WebKit/LayoutTests/typedcssom/cssPerspective.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

test(function() {
  var calcLength = new CSSCalcValue({px: 10, percent: 3.2});
  assert_throws(null, function() { new CSSPerspective(calcLength) });
}, "Constructor should throw an error for CSSCalcValues with a percentage type");

test(function() {
  var simpleLength = new CSSUnitValue(10, 'percent');
  assert_throws(new TypeError(), function() { new CSSPerspective(simpleLength) });
}, "Constructor should throw an error for CSSUnitValues with a percentage type");

test(function() {
  var simpleLength = new CSSUnitValue(10, 'px');
  var calcLength = new CSSCalcValue({px: 10, em: 3.2});
  var perspectiveTransformSimple = new CSSPerspective(simpleLength);
  var perspectiveTransformCalc = new CSSPerspective(calcLength);

  assert_equals(perspectiveTransformSimple.toString(), 'perspective(10px)');
  assert_equals(perspectiveTransformCalc.toString(), 'perspective(calc(3.2em + 10px))');
}, "toString should return a string of form perspective(<CSSLengthValue.cssString()>)");

</script>

<body>
</body>
